reorganized ngrx files
