+++ client/src/app/shared/state/display-state/layout.state.ts
@@ -1,0 +1,6 @@
+import { EntityState } from '@ngrx/entity';
+import { Layout } from '../../models/layout.model';
+
+export interface State extends EntityState<Layout> {
+  selectedLayoutId: number | string | null;
+}
