+++ client/src/app/shared/state/index.ts
@@ -1,0 +1,7 @@
+import { AppState } from "./app.state";
+import { reducers } from "./app.state";
+
+export {
+    AppState,
+    reducers,
+}
